Corecciones para hacer funcional el fronted/backend

+++ backend/src/products/commands/deleteProduct.ts
@@ -2,23 +2,27 @@ import {PrismaClient} from "@prisma/client";
 import {Request, Response} from "express";
 const prisma = new PrismaClient();
 
-export function deleteProduct(
+export async function deleteProduct(
   req: Request<{id: number}, {}, {}>,
   res: Response<{} | {error: string}>
 ) {
   try {
-    prisma.product.delete({
+    const id = Number(req.params.id);
+    if (isNaN(id)) {
+      res.status(400).json({error: "Invalid product id"});
+      return;
+    }
+    await prisma.product.delete({
       where: {
-        id: req.params.id,
+        id: id,
       },
     });
     res.status(204).send();
-    console.log(`Product with id ${req.params.id} deleted`);
-    prisma.$disconnect();
+    await prisma.$disconnect();
   } catch (error) {
     res
       .status(500)
-      .json({error: "An error occurred while deleting the product"});
-    prisma.$disconnect();
+      .json({error: "An error occurred while deleting the product: " + error});
+    await prisma.$disconnect();
   }
 }
